ai_resume_builder/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
jinja2==3.1.2
httpx==0.25.2

# AI/ML
google-generativeai>=0.8.0
sentence-transformers>=2.2.2
scikit-learn==1.3.2
pandas==2.1.3

# PDF Generation
reportlab==4.0.7

# Monitoring & Observability
prometheus-client==0.19.0

# Caching
redis==5.0.1

# Testing
pytest==7.4.3
